+++ l2geth/eth/config.go
@@ -81,7 +81,8 @@ var DefaultConfig = Config{
 		// is additional overhead that is unaccounted. Round down to 127000 for
 		// safety.
 		MaxCallDataSize: 127000,
-		L1GasPrice:      big.NewInt(100 * params.GWei),
+		DataPrice:       big.NewInt(100 * params.GWei),
+		ExecutionPrice:  big.NewInt(0),
 	},
 	DiffDbCache: 256,
 }