+++ packages/data-transport-layer/src/services/l2-ingestion/service.ts
@@ -1,5 +1,5 @@
 /* Imports: External */
-import { BaseService } from '@eth-optimism/core-utils'
+import { BaseService } from '@eth-optimism/common-ts'
 import { JsonRpcProvider } from '@ethersproject/providers'
 import { BigNumber } from 'ethers'
 import { LevelUp } from 'levelup'