+++ packages/data-transport-layer/src/services/main/service.ts
@@ -1,5 +1,5 @@
 /* Imports: External */
-import { BaseService } from '@eth-optimism/core-utils'
+import { BaseService } from '@eth-optimism/common-ts'
 import { LevelUp } from 'levelup'
 import level from 'level'